app/requirements.txt
black~=22.0
flake8>=3.7.9, <5.0
ipython>=7.31.1, <8.0; python_version < '3.8'
ipython~=8.10; python_version >= '3.8'
isort~=5.0
jupyter~=1.0
jupyterlab_server>=2.11.1, <2.16.0
jupyterlab~=3.0
kedro==0.18.13
kedro-datasets~=1.0
pyarrow~=13.0
fastparquet~=2023.8
openpyxl~=3.1
kedro-telemetry~=0.2.0
kedro-viz~=6.0
nbstripout~=0.4
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=7.2
scikit-learn~=1.0
adlfs==2023.9.0
hyperopt~=0.2.7
interpret~=0.4.4
interpret_core==0.4.4
matplotlib==3.7.2
numpy==1.24.4
optuna==3.4.0
pandas==2.1.3
Pillow==10.0.0
plotly==5.16.1
PyYAML==6.0.1
scikit_learn==1.3.2
seaborn==0.13.0
streamlit==1.27.2
tqdm==4.66.1